requirements.txt
PyYAML==5.3.1
selenium==3.141.0
urllib3==1.25.10

qbittorrent~=0.1.6